+++ actions/subtasks.ts
@@ -3,10 +3,11 @@ import { subtasks } from '@/db/schema'
 import { db } from '@/db'
 import { eq } from 'drizzle-orm'
 import { revalidateTag } from 'next/cache'
+import { NewSubtasks, Subtasks } from '@/types'
 
-export const deleteSubtasks = async (subtasksArr: any) => {
+export const deleteSubtasks = async (subtasksArr: Subtasks[]) => {
   try {
-    subtasksArr.forEach(async (sub: any) => {
+    subtasksArr.forEach(async (sub: Subtasks) => {
       await db.delete(subtasks).where(eq(subtasks.id, sub.id))
     })
   } catch (error) {
@@ -14,7 +15,10 @@ export const deleteSubtasks = async (subtasksArr: any) => {
   }
 }
 
-export const updateSubtasks = async (subtasksToUpdate: any, taskId: string) => {
+export const updateSubtasks = async (
+  subtasksToUpdate: Subtasks[],
+  taskId: string
+) => {
   try {
     const savedSubtasks = await db.query.subtasks.findMany({
       where: eq(subtasks.taskId, taskId),
@@ -22,19 +26,22 @@ export const updateSubtasks = async (subtasksToUpdate: any, taskId: string) => {
 
     const subtasksToDelete = savedSubtasks.filter(
       (sst) =>
-        sst.taskId && !subtasksToUpdate.some((ssu: any) => ssu.id === sst.id)
+        sst.taskId &&
+        !subtasksToUpdate.some((ssu: Subtasks) => ssu.id === sst.id)
     )
 
-    subtasksToUpdate.forEach(async (sub: any) => {
+    subtasksToUpdate.forEach(async (sub: Subtasks) => {
       const isIncluded = savedSubtasks.some((ss) => ss.id === sub.id)
 
       if (isIncluded) {
+        const subId = sub.id!
+
         await db
           .update(subtasks)
           .set({
             title: sub.title,
           })
-          .where(eq(subtasks.id, sub.id))
+          .where(eq(subtasks.id, subId))
       } else {
         await db.insert(subtasks).values({
           taskId: taskId,
@@ -43,17 +50,19 @@ export const updateSubtasks = async (subtasksToUpdate: any, taskId: string) => {
       }
     })
 
-    await deleteSubtasks(subtasksToDelete)
+    if (subtasksToDelete && subtasksToDelete.length !== 0) {
+      await deleteSubtasks(subtasksToDelete)
+    }
   } catch (error) {
     console.log(error)
   }
 }
 
-export const completeSubtask = async (id: string, bool: boolean) => {
+export const completeSubtask = async (id: string, value: boolean) => {
   try {
     await db
       .update(subtasks)
-      .set({ isCompleted: bool })
+      .set({ isCompleted: value })
       .where(eq(subtasks.id, id))
 
     revalidateTag('dashboard:boardSelected')
@@ -62,7 +71,10 @@ export const completeSubtask = async (id: string, bool: boolean) => {
   }
 }
 
-export const createSubtasks = async (subtasksArr: any, taskId: string) => {
+export const createSubtasks = async (
+  subtasksArr: NewSubtasks[],
+  taskId: string
+) => {
   try {
     subtasksArr.forEach(async ({ title }: { title: string }) => {
       await db.insert(subtasks).values({
